% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybinom.R
\name{mybinom}
\alias{mybinom}
\title{Customized Binomial Distribution}
\usage{
mybinom(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{A vector of the number of observations}

\item{n}{A vector of values from 1 to n}

\item{p}{A vector of probabilities}
}
\value{
A plot and a table of binomial distribution
}
\description{
Customized Binomial Distribution
}
\examples{
\dontrun{mybinom(iter=1000, n=15, p=0.6)}
}
