% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap sample statistics}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{columns}

\item{x}{a vector of the random sample}

\item{fun}{function type, i.e. quantile, IQR, sd, var, mean, median}

\item{alpha}{the significance level used to compute the confidence level}

\item{cx}{character expansion to adjust the size of text}

\item{...}{}
}
\description{
Bootstrap sample statistics
}
\examples{
\dontrun{myboot2(iter=10000,x=sample, fun="mean", alpha=0.05, xlab="Mean", col="Green")}
}
