% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Find confidence interval}
\usage{
myci(data, alpha = 0.05)
}
\arguments{
\item{data}{a data set}

\item{alpha}{default value is 0.05}
}
\value{
a confidence interval
}
\description{
Find confidence interval
}
\examples{
\dontrun{myci(d)}
}
