% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{Central Limit Theorem in the Uniform Distribution
It will approach a normal distribution as the sample size grows.}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{rows}

\item{iter}{columns}

\item{a}{lower limit}

\item{b}{upper limit}
}
\value{
a histogram of the distribution of the sample mean.
}
\description{
Central Limit Theorem in the Uniform Distribution
It will approach a normal distribution as the sample size grows.
}
\examples{
\dontrun{mycltu(n=20,iter=100000)}
}
