% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Maximum likelihood Estimate function}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{Function type}

\item{x}{a vector of x}

\item{param}{a vector of parameter}

\item{...}{}
}
\description{
Maximum likelihood Estimate function
}
