% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Title    Normal Distribution Curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{The mean value}

\item{sigma}{The standard deviation}

\item{a}{The limit on the right x}
}
\value{
The shaded area of probability P(X<=a)
}
\description{
Title    Normal Distribution Curve
}
\examples{
\dontrun{myncurve(mu=10,sigma=5,a=6)}
}
