% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{Display P-value areas}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{A data set}

\item{xmax}{A maximum value of x}

\item{n}{A sample size}

\item{alpha}{A significant level}
}
\value{
A plot of p-value
}
\description{
Display P-value areas
}
\examples{
\dontrun{mypvalue(tcalc,n=30,alpha=0.05)}
}
