% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Title    Calculate the number of tickets to be sold}
\usage{
ntickets(N = 200, gamma = 0.02, p = 0.95)
}
\arguments{
\item{N}{the number of seats in the flight}

\item{gamma}{the probability a plane will be truly overbooked}

\item{p}{the probability of a "show"}
}
\value{
display a named list containing nd, nc, N, p and gamma; and
two plots for the discrete and for the continuous case to find the optimal tickets sold
}
\description{
Title    Calculate the number of tickets to be sold
}
\examples{
\dontrun{ntickets(N=400,gamma = 0.02, p = 0.95)}
}
